refactor: restructure files in domains
