+++ src/indexer/abis/v1/ProjectRegistry.json
