+++ src/indexer/abis/v1/QuadraticFundingVotingStrategyFactory.json
