+++ src/indexer/abis/v1/QuadraticFundingVotingStrategyImplementation.json
