+++ src/indexer/abis/v2/ProjectRegistry.json
