+++ src/indexer/abis/v2/QuadraticFundingVotingStrategyFactory.json
