+++ src/indexer/abis/v2/QuadraticFundingVotingStrategyImplementation.json
