+++ src/indexer/handlers/roundMetaPtrUpdated.ts
@@ -1,7 +1,7 @@
 import { Indexer, JsonStorage, Event } from "chainsauce";
-import { fetchJsonCached as ipfs } from "../ipfs.js";
+import { fetchJsonCached as ipfs } from "../../utils/ipfs.js";
 
-export default async function applicationMetaPtrUpdated(
+export default async function roundMetaPtrUpdated(
   { cache, storage: db }: Indexer<JsonStorage>,
   event: Event
 ) {
@@ -9,7 +9,7 @@ export default async function applicationMetaPtrUpdated(
 
   await db.collection("rounds").updateById(id, (round) => ({
     ...round,
-    applicationMetaPtr: event.args.newMetaPtr.pointer,
+    metaPtr: event.args.newMetaPtr.pointer,
   }));
 
   return async () => {
@@ -18,8 +18,8 @@ export default async function applicationMetaPtrUpdated(
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     await db.collection("rounds").updateById(id, (round: any) => {
-      if (round.applicationMetaPtr === event.args.newMetaPtr.pointer) {
-        return { ...round, applicationMetadata: metadata };
+      if (round.metaPtr === event.args.newMetaPtr.pointer) {
+        return { ...round, metadata };
       }
 
       return round;
