+++ src/utils/ipfs.ts
@@ -1,7 +1,7 @@
 import { Cache } from "chainsauce";
 
 import fetchRetry from "./fetchRetry.js";
-import config from "./config.js";
+import config from "../config.js";
 
 export async function fetchJsonCached<T>(
   cid: string,
